Navigate Between Pages
Client-Side History Support

When you hit the Back button, it navigates the page to the index page entirely
via the client; next/link does all the handling for you.

You don't need to write even a single line of client-side routing code.
Simply link pages; it just works!

Link is Just a Higher Order Component (HOC)
Actually, the style prop on next/link has no effect. That's because next/link
is just a which only accepts the "href" and some similar props. If you need to
style it, you need to do it to the underlying component.

In this case, it's our anchor.

Link Works With Anything
Just like a button, you can place any of your custom React components or even a
div within a Link.

The only requirement for components placed inside a Link is they should accept
an onClick prop.