Server Side Support for Clean URLs
Our /post page accepts the title via the query string parameter title. In
client side routing, we can easily give it a proper value with URL masking.
(via the as prop in Link).

But in the server route, we don't have that title because we only have an ID
for the blog post in the URL. So, in that case, we set the ID as the server
side query string param.