Fetching Data for Pages
That's a static async function you can add into any page in your app. Using
that, we can fetch data and send them as props to our page.

As you can see, now we are fetching Batman TV shows and passing them into our
page as the 'shows' prop.

As you can see in the above getInitialProps function, it prints the number of data fetched to the console.

That's because we render the page on the server. So, we already have the data
and there is no reason to fetch it again in the client.