Fetching Data for Pages
Here we can only see the message on the browser console.

That's because we navigated to the post page via the client side. Then fetching
data from the client side is the best way to do it.

Now you have learned one of the most crucial features of Next.js that makes it
ideal for universal data fetching and server side rendering.

We have learned the basics of getInitialProps which is enough for most of the
use cases.