Lazy Loading Modules
We use the firebase module, only when the user is trying to navigate into a
different page. So, if we can load the firebase module at that time, that is a
huge win for our app.

Luckily, we can easily do that with Next.js's  functionality.

Let's do it.

The firebase related code is located in the lib/load-db.js file. Here is the
content:

Test Result
As you have witnessed, it only loads when you navigate a page for the first
time. Here is what is actually happening.

At the first time, getIntitialProps of the pages/post.js page imports the
firebase module (via lib/load-db.js). So, the app loads the bundle.

Even the second time, pages/index.js page imports the firebase module. But at
that time, it is already loaded and there is no reason to load it again.

Finally
Frankly, this example is not the perfect use case for lazy loading. That is
simply because,

You need the firebase module in all of the pages.
Lazy loaded firebase module reduces the size of the main JavaScript bundle
app.js, but it doesn't affect the initial page loading time since the page is
server rendered.
Loading of the main JavaScript bundle doesn't block the initial HTML rendering
The only benefit this gives us, is the quick JavaScript interactivity because
the app.js loads faster due to the reduced size.

Anyway, this is a pretty good example where we can demonstrate the lazy loading
module.  So, you can use it in your app when it is needed.