Lazy Loading Components
We use a lot of React components in our applications. We include them in our
main JavaScript bundle (app.js) or in one of the page's JavaScript bundles.
Sometimes some of these components are pretty big.

Since it's used in almost all of the pages, react-highlight is included in our
main JavaScript bundle.

But we don't need to use the syntax highlighter on every page. It's only needed
when there's a code sample in the blog post.

So, if we can load react-highlight only when needed, that will reduce the
initial app bundle dramatically, and that will help our app load faster.

This is something we can do with Next.js (since version 3.0).
Let's get started.